Refactoring architecture
