+++ Application/src/components/AppTemplate.tsx
