--- Application/src/environment/Global_variable.tsx
@@ -1,5 +1,0 @@
-enum GLOBAL {
-    version = "0.1",
-}
-
-export default GLOBAL;
