+++ Application/src/models/User.tsx
@@ -1,0 +1,5 @@
+export default interface User {
+    id: number;
+    username: string;
+    password: string;
+  }
